Reorganize project layout with workspaces
Workspaces are a feature in Cargo that make it easier to manage
muti-crate projects. The project will now have a single `target`
directory instead of having it both in `bs_lib` and `bs_bin`. This means
that dependencies will only be compiled once and not duplicated. This
also lets you build and lint the code from the root of the repository
instead of having to handle one of the crates at a time.